requirements.txt
infrahouse-core ~= 0.17
python-mimeparse ~=2.0
pytest-infrahouse ~= 0.11
